refactor: consolidate logic
