+++ crates/cdk/src/wallet/swap/saga/mod.rs
@@ -23,7 +23,6 @@ use cdk_common::wallet::{
 };
 use tracing::instrument;
 
-use self::compensation::RevertSwapProofReservation;
 use self::state::{Finalized, Initial, Prepared};
 use crate::amount::SplitTarget;
 use crate::dhke::construct_proofs;
@@ -32,10 +31,10 @@ use crate::nuts::{nut10, Proofs, SpendingConditions, State};
 use crate::types::ProofInfo;
 use crate::wallet::saga::{
     add_compensation, clear_compensations, execute_compensations, new_compensations, Compensations,
+    RevertProofReservation as RevertSwapProofReservation,
 };
 use crate::{Amount, Error, Wallet};
 
-pub mod compensation;
 pub mod resume;
 pub mod state;
